Registro de notas de credito

+++ src/main/java/com/ccadmin/app/shared/model/entity/BusinessConfigEntity.java
@@ -14,11 +14,11 @@ import java.math.BigDecimal;
 @IdClass(BusinessConfigEntityID.class)
 public class BusinessConfigEntity implements Serializable {
 
-    public int GroupId;
+    public Integer GroupId;
     @Id
     public String GroupCod;
     @Id
-    public int ConfigCorr;
+    public Integer ConfigCorr;
     public String ConfigCod;
     public String ConfigVal;
     public String ConfigName;
@@ -27,11 +27,11 @@ public class BusinessConfigEntity implements Serializable {
     public String Str2Config;
     public String Str3Config;
     public String Str4Config;
-    public int Num1Config;
-    public int Num2Config;
-    public int Num3Config;
-    public int Num4Config;
-    public float Dcm1Config;
+    public Integer Num1Config;
+    public Integer Num2Config;
+    public Integer Num3Config;
+    public Integer Num4Config;
+    public BigDecimal Dcm1Config;
     public BigDecimal Dcm2Config;
     public BigDecimal Dcm3Config;
     public BigDecimal Dcm4Config;
